changed getById in interface DAO from int to long to former convention

+++ src/main/java/com/project/db/dao/DAO.java
@@ -24,5 +24,5 @@ public interface DAO <T>{
     public void delete(T u);
 
     @Nullable
-    public T getById(@NotNull int id);
+    public T getById(@NotNull long id);
 }
